move and rename text
