+++ src/shapes/Text/TextSVGExportMixin.ts
@@ -1,12 +1,12 @@
 // @ts-nocheck
-import { config } from '../config';
-import { TSVGReviver } from '../typedefs';
-import { escapeXml } from '../util/lang_string';
-import { createSVGRect } from '../util/misc/svgParsing';
-import { hasStyleChanged } from '../util/misc/textStyles';
-import { toFixed } from '../util/misc/toFixed';
-import { FabricObjectSVGExportMixin } from './object.svg_export';
-import type { TextStyleDeclaration } from './text_style.mixin';
+import { config } from '../../config';
+import { TSVGReviver } from '../../typedefs';
+import { escapeXml } from '../../util/lang_string';
+import { createSVGRect } from '../../util/misc/svgParsing';
+import { hasStyleChanged } from '../../util/misc/textStyles';
+import { toFixed } from '../../util/misc/toFixed';
+import { FabricObjectSVGExportMixin } from '../../mixins/object.svg_export';
+import type { TextStyleDeclaration } from './StyledText';
 
 const multipleSpacesRegex = /  +/g;
 const dblQuoteRegex = /"/g;
