+++ src/util/types.ts
@@ -6,7 +6,7 @@ import type {
 } from '../shapes/Object/Object';
 import type { FabricObjectWithDragSupport } from '../shapes/Object/InteractiveObject';
 import type { TFiller } from '../typedefs';
-import type { Text } from '../shapes/text.class';
+import type { Text } from '../shapes/Text/Text';
 import type { Pattern } from '../pattern.class';
 import type { IText } from '../shapes/IText/IText';
 import type { Textbox } from '../shapes/textbox.class';
